requirements.txt
streamlit
python-dotenv
pydantic
boto3